+++ flower.core/src/test/java/com/ly/train/flower/core/filter/OpentracingFilterTest.java
@@ -37,7 +37,41 @@ public class OpentracingFilterTest extends TestBase {
     serviceFlow.buildFlow(UserServiceA.class, UserServiceB.class);
     serviceFlow.buildFlow(UserServiceB.class, UserServiceC1.class);
     serviceFlow.buildFlow(UserServiceC1.class, OpenTracerService.class);
-    serviceFlow.setFilters(Sets.newSet("mockOne", "mockTwo"));
+    serviceFlow.setFilters(Sets.newSet("mockOneFilter", "mockTwoFilter"));
+    serviceFlow.build();
+    final FlowRouter router = serviceFacade.buildFlowRouter(flowName, 2 << 3);
+
+    User user = new User();
+    user.setName("响应式编程 ");
+    user.setAge(2);
+    Object o = router.syncCallService(user);
+    System.out.println("响应结果： " + o);
+  }
+
+  @Test
+  public void testCounterFilter() throws Exception {
+    ServiceFlow serviceFlow = serviceFactory.getOrCreateServiceFlow(flowName);
+    serviceFlow.buildFlow(UserServiceA.class, UserServiceB.class);
+    serviceFlow.buildFlow(UserServiceB.class, UserServiceC1.class);
+    serviceFlow.buildFlow(UserServiceC1.class, OpenTracerService.class);
+    serviceFlow.setFilters(Sets.newSet("mockOneFilter", "mockTwoFilter", "counterFilter"));
+    serviceFlow.build();
+    final FlowRouter router = serviceFacade.buildFlowRouter(flowName, 2 << 3);
+
+    User user = new User();
+    user.setName("响应式编程 ");
+    user.setAge(2);
+    Object o = router.syncCallService(user);
+    System.out.println("响应结果： " + o);
+  }
+
+  @Test
+  public void testAccessLogFilter() throws Exception {
+    ServiceFlow serviceFlow = serviceFactory.getOrCreateServiceFlow(flowName);
+    serviceFlow.buildFlow(UserServiceA.class, UserServiceB.class);
+    serviceFlow.buildFlow(UserServiceB.class, UserServiceC1.class);
+    serviceFlow.buildFlow(UserServiceC1.class, OpenTracerService.class);
+    serviceFlow.setFilters(Sets.newSet("mockOneFilter", "mockTwoFilter", "accessLogFilter"));
     serviceFlow.build();
     final FlowRouter router = serviceFacade.buildFlowRouter(flowName, 2 << 3);
