+++ flower.core/src/test/resources/META-INF/services/flower/com.ly.train.flower.filter.Filter
@@ -1,2 +1,2 @@
-mockOne=com.ly.train.flower.core.filter.impl.MockOneFilter
-mockTwo=com.ly.train.flower.core.filter.impl.MockTwoFilter
+mockOneFilter=com.ly.train.flower.core.filter.impl.MockOneFilter
+mockTwoFilter=com.ly.train.flower.core.filter.impl.MockTwoFilter
