+++ flower.filter/flower.filter.api/src/main/java/com/ly/train/flower/filter/FilterConfig.java
@@ -1,0 +1,8 @@
+package com.ly.train.flower.filter;
+
+/**
+ * @author leeyazhou
+ */
+public class FilterConfig {
+
+}
