+++ flower.filter/flower.filter.api/src/main/java/com/ly/train/flower/filter/impl/TimeoutFilter.java
@@ -13,28 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * 
- */
-package com.ly.train.flower.filter;
+package com.ly.train.flower.filter.impl;
 
+import com.ly.train.flower.common.core.service.ServiceContext;
 import com.ly.train.flower.common.logging.Logger;
 import com.ly.train.flower.common.logging.LoggerFactory;
+import com.ly.train.flower.filter.AbstractFilter;
+import com.ly.train.flower.filter.FilterChain;
 
 /**
  * @author leeyazhou
  * 
  */
-public abstract class AbstractFilter implements Filter {
-  static final Logger logger = LoggerFactory.getLogger(AbstractFilter.class);
+public class TimeoutFilter extends AbstractFilter {
+  private static final Logger logger = LoggerFactory.getLogger(TimeoutFilter.class);
 
   @Override
-  public void init(FilterConfig filterConfig) {}
-
-  @Override
-  public int getOrder() {
-    return 0;
+  public Object doFilter(Object message, ServiceContext context, FilterChain chain) {
+    logger.info("message : {}, context : {}", message, context);
+    return chain.doFilter(message, context);
   }
-
 
 }
