+++ flower.filter/flower.filter.opentracing/src/main/resources/META-INF/services/flower/com.ly.train.flower.filter.Filter
@@ -1,1 +1,1 @@
-opentracer=com.ly.train.flower.filter.impl.OpenTracingFilter
+opentracingFilter=com.ly.train.flower.filter.impl.OpenTracingFilter
